backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
stripe==7.0.0
web3==6.0.0
pillow==11.2.1
django-storages==1.14.0
drf-spectacular==0.26.5
gunicorn==21.2.0
django-widget-tweaks==1.5.0
django-filter==23.5
whitenoise==6.6.0